Refactor waiter helpers into a separate package
